Update server to new version and start working on gallery backend

+++ server/src/models/recipe.rs
@@ -2,7 +2,7 @@ use serde_derive::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct Recipe {
-    pub id: i32,
+    pub id: i64,
     pub title: String,
     pub description: String,
     pub content: String,
@@ -10,7 +10,7 @@ pub struct Recipe {
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct RecipeListItem {
-    pub id: i32,
+    pub id: i64,
     pub title: String,
     pub description: String,
 }
